pyproject.toml
[tool.poetry]
name = "online-school"
version = "0.1.0"
description = ""
authors = ["dimapb"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
ipython = "^8.28.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
redis = "^5.1.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
